+++ src/tests/backend_mock.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock
 from argparse import Namespace
 from tmtccmd.core.backend import TmTcHandler
 from tmtccmd.config.hook_base import TmTcHookBase
-from tmtccmd.core.definitions import CoreComInterfaces, CoreModeList, CoreServiceList
+from tmtccmd.config.definitions import CoreComInterfaces, CoreModeList, CoreServiceList
 from tmtccmd.core.frontend_base import FrontendBase
