+++ src/tmtccmd/com_if/serial_com_if.py
@@ -15,7 +15,6 @@ import serial.tools.list_ports
 
 from tmtccmd.ecss.tc import PusTelecommand
 from tmtccmd.com_if.com_interface_base import CommunicationInterface
-from tmtccmd.core.definitions import CoreComInterfaces
 from tmtccmd.utility.tmtc_printer import TmTcPrinter
 from tmtccmd.pus_tm.factory import PusTelemetryFactory, PusTmListT
 from tmtccmd.utility.tmtcc_logger import get_logger
