+++ src/tmtccmd/com_if/tcpip_utilities.py
@@ -4,7 +4,7 @@ import struct
 import enum
 from typing import Union
 
-from tmtccmd.core.definitions import ethernet_address_t
+from tmtccmd.config.definitions import ethernet_address_t
 from tmtccmd.utility.json_handler import check_json_file
 from tmtccmd.utility.tmtcc_logger import get_logger
 from tmtccmd.utility.json_handler import JsonKeyNames
