+++ src/tmtccmd/config/definitions.py
