+++ src/tmtccmd/defaults/tc_packing.py
@@ -4,7 +4,7 @@ from collections import deque
 from tmtccmd.ecss.tc import PusTelecommand
 
 from tmtccmd.pus_tc.definitions import TcQueueT
-from tmtccmd.core.definitions import CoreServiceList
+from tmtccmd.config.definitions import CoreServiceList
 from tmtccmd.pus_tc.service_17_test import pack_service17_ping_command
 from tmtccmd.pus_tc.service_5_event import pack_generic_service5_test_into
 from tmtccmd.utility.tmtcc_logger import get_logger
