+++ src/tmtccmd/pus_tc/definitions.py
@@ -1,5 +1,5 @@
 from typing import Tuple, Union, Deque
-from tmtccmd.core.definitions import QueueCommands
+from tmtccmd.config.definitions import QueueCommands
 from tmtccmd.ecss.tc import PusTcTupleT, PusTelecommand
 
 TcAuxiliaryTupleT = Tuple[QueueCommands, any]
