+++ src/tmtccmd/pus_tc/service_17_test.py
@@ -1,6 +1,4 @@
-import enum
-
-from tmtccmd.core.definitions import QueueCommands
+from tmtccmd.config.definitions import QueueCommands
 from tmtccmd.pus_tc.definitions import PusTelecommand, TcQueueT
 from tmtccmd.config.globals import get_global_apid
 from tmtccmd.pus.service_17_test import Srv17Subservices
