+++ src/tmtccmd/sendreceive/cmd_sender_receiver.py
@@ -14,7 +14,7 @@ if the first reply has not been received.
 import time
 
 from tmtccmd.com_if.com_interface_base import CommunicationInterface
-from tmtccmd.core.definitions import QueueCommands, CoreGlobalIds
+from tmtccmd.config.definitions import QueueCommands, CoreGlobalIds
 from tmtccmd.utility.tmtc_printer import TmTcPrinter
 from tmtccmd.utility.tmtcc_logger import get_logger
