+++ src/tmtccmd/sendreceive/multiple_cmds_sender_receiver.py
@@ -7,7 +7,7 @@ import time
 from typing import Union, Deque
 from collections import deque
 
-from tmtccmd.core.definitions import CoreGlobalIds
+from tmtccmd.config.definitions import CoreGlobalIds
 from tmtccmd.sendreceive.sequential_sender_receiver import SequentialCommandSenderReceiver
 from tmtccmd.com_if.com_interface_base import CommunicationInterface
 from tmtccmd.utility.tmtc_printer import TmTcPrinter
